Added tab autocomplete
A horribly long function for autocompleting commands or files, this
will have to be refactored at some point, but for now it is just a
nice to have feature.

At the first word of a console input will be treated as a command, the
rest as files. One tab will fill in a unique match or the common
characters from the candidates, if nothing is added a second tab will
list the available candidates.